chore(java): :recycle: minor refactoring done for few methods
