+++ core-java/src/main/java/com/github/wasiqb/boyka/config/FrameworkSetting.java
@@ -41,7 +41,6 @@ public class FrameworkSetting {
     private static final Logger LOGGER = getLogger ();
 
     private Map<String, ApiSetting> api;
-    private LogSetting              logs = new LogSetting ();
     private UISetting               ui;
 
     /**
