+++ core-java/src/main/java/com/github/wasiqb/boyka/config/api/LogSetting.java
@@ -19,19 +19,14 @@ package com.github.wasiqb.boyka.config.api;
 import lombok.Data;
 
 /**
- * API setting class.
+ * Appium server log settings.
  *
  * @author Wasiq Bhamla
- * @since 17-Feb-2022
+ * @since 07-Sept-2022
  */
 @Data
-public class ApiSetting {
-    private String     basePath          = "";
-    private String     baseUri;
-    private int        connectionTimeout = 5;
-    private LogSetting logging           = new LogSetting ();
-    private int        port;
-    private int        readTimeout       = 5;
-    private String     schemaPath        = "";
-    private int        writeTimeout      = 5;
+public class LogSetting {
+    private boolean enable   = true;
+    private boolean request  = true;
+    private boolean response = true;
 }
