+++ core-java/src/main/java/com/github/wasiqb/boyka/config/ui/UISetting.java
@@ -30,6 +30,7 @@ import lombok.Data;
  */
 @Data
 public class UISetting {
+    private LogSetting                 logging    = new LogSetting ();
     private Map<String, MobileSetting> mobile;
     private ScreenshotSetting          screenshot = new ScreenshotSetting ();
     private TimeoutSetting             timeout    = new TimeoutSetting ();
